refactor: change module name from time to cli
