+++ craft_application/util/__init__.py
@@ -36,7 +36,7 @@ from craft_application.util.snap_config import (
 from craft_application.util.string import humanize_list, strtobool
 from craft_application.util.system import get_parallel_build_count
 from craft_application.util.yaml import dump_yaml, safe_yaml_load
-from craft_application.util.time import format_timestamp
+from craft_application.util.cli import format_timestamp
 
 __all__ = [
     "get_unique_callbacks",
