+++ tests/unit/util/test_cli.py
@@ -14,25 +14,28 @@
 #  You should have received a copy of the GNU Lesser General Public License
 #  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Utilities to deal with time."""
+from datetime import datetime, timedelta, timezone
 
-import datetime
+import pytest
+from craft_application.util import format_timestamp
 
 
-def format_timestamp(dt: datetime.datetime) -> str:
-    """Convert a datetime object (with or without timezone) to a string.
-
-    The format is
-
-        <DATE>T<TIME>Z
-
-    Always in UTC.
-    Borrowed from charmcraft.
-    """
-    if dt.tzinfo is not None and dt.tzinfo.utcoffset(None) is not None:
-        # timezone aware
-        dtz = dt.astimezone(datetime.timezone.utc)
-    else:
-        # timezone naive, assume it's UTC
-        dtz = dt
-    return dtz.strftime("%Y-%m-%dT%H:%M:%SZ")
+@pytest.mark.parametrize(
+    ("dt_obj", "expected"),
+    [
+        (
+            datetime(2024, 5, 23, 13, 24, 0),
+            "2024-05-23T13:24:00Z",
+        ),
+        (
+            datetime(2024, 5, 23, 13, 24, 0, tzinfo=timezone.utc),
+            "2024-05-23T13:24:00Z",
+        ),
+        (
+            datetime(2024, 5, 23, 13, 24, 0, tzinfo=timezone(timedelta(hours=-5))),
+            "2024-05-23T18:24:00Z",
+        ),
+    ],
+)
+def test_timezone_parsing(dt_obj: datetime, expected: str) -> None:
+    assert format_timestamp(dt=dt_obj) == expected
